Moved to amazon db
1. Tested adding cards, goals, assists and matches.
2. NEED to test CRUD of players, seasons, opponents, users, etc.

+++ QuilmesAC/Controllers/OpponentController.cs
@@ -124,7 +124,7 @@
         }
 
         /// <summary> Modal delete </summary>
-        public ContentResult DeleteOpponent(long id)
+        public ContentResult DeleteOpponent(int id)
         {
             var opponent = QuilmesModel.GetOpponentByID(id);
             QuilmesModel.Delete(opponent);
